+++ MoneyMaker/Services/PortfolioService.cs
@@ -38,7 +38,7 @@ namespace MoneyMaker.Services
         }
 
 
-        public async Task<PortfolioEntry> GetPortfolio(string UserId, string EntryCurrencySym, int EntryValue)
+        public async Task<PortfolioEntry> GetPortfolio(string UserId, string EntryCurrencySym, float EntryValue)
         {
             object[] param = { UserId, EntryCurrencySym, EntryValue };
             PortfolioEntry port = await _context.PortfolioEntry.FindAsync(param);
@@ -78,7 +78,7 @@ namespace MoneyMaker.Services
             return port;
         }
 
-        public async Task<Boolean> DeletePortfolio(string UserId, string EntryCurrencySym, int EntryValue)
+        public async Task<Boolean> DeletePortfolio(string UserId, string EntryCurrencySym, float EntryValue)
         {
             object[] param = { UserId, EntryCurrencySym, EntryValue };
             PortfolioEntry port = await _context.PortfolioEntry.FindAsync(param);
@@ -92,7 +92,7 @@ namespace MoneyMaker.Services
             return false;
         }
 
-        public async Task<bool> PortfolioExists(string UserId, string EntryCurrencySym, int EntryValue)
+        public async Task<bool> PortfolioExists(string UserId, string EntryCurrencySym, float EntryValue)
         {
             return _context.PortfolioEntry.Any(p => (p.UserId == UserId && p.EntryCurrencySym == EntryCurrencySym && p.EntryCurrencySym == EntryCurrencySym));
         }
